feat: add metadata event, update init event
VOD vs LIVE is not known until the manifest has been parsed.
It can also change during a session. Introduce a metadata
event to carry non-QoS metadata related to the session.

The payload can be extended with any property that the client
and backend can handle, with the most common cases included in
the standard.